Refactor: removed model service layer

+++ Backend/API/Data/Repositories/AircraftRepository.cs
@@ -1,29 +1,34 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using API.Domain.Models;
-using API.Domain.Repositories;
-using API.Domain.Services.Communication;
-using API.Persistance.Contexts;
+using API.Domain.Interfaces;
+using API.Data.Contexts;
 using Microsoft.EntityFrameworkCore;
+using System;
 
-namespace API.Persistance.Repositories
+namespace API.Data.Repositories
 {
     public class AircraftRepository : BaseRepository, IAircraftRepository
     {
         public AircraftRepository(AppDbContext context) : base(context) { }
 
+        // TODO: error handling
         public async Task<IEnumerable<Aircraft>> GetAllAircraftsAsync()
         {
             return await _context.Aircrafts.ToListAsync();
         }
 
-        public async Task<Aircraft> FindAircraftByIdAsync(int id)
+        public async Task<Aircraft> GetAircraftByIdAsync(int id)
         {
             return await _context.Aircrafts.FindAsync(id);
         }
 
-        public async Task AddAircraftAsync(Aircraft aircraft)
+        public async Task CreateAircraftAsync(Aircraft aircraft)
         {
+            if (aircraft == null)
+            {
+                throw new ArgumentNullException(nameof(aircraft));
+            }
             await _context.Aircrafts.AddAsync(aircraft);
         }
 
@@ -32,8 +37,12 @@ namespace API.Persistance.Repositories
             _context.Aircrafts.Update(aircraft);
         }
 
-        public void RemoveAircraft(Aircraft aircraft)
+        public void DeleteAircraft(Aircraft aircraft)
         {
+            if (aircraft == null) 
+            {
+                throw new ArgumentNullException(nameof(aircraft));
+            }
             _context.Aircrafts.Remove(aircraft);
         }
     }
